requirements.txt
google-cloud-storage
google-cloud-bigquery
functions-framework==3.*
jsonschema
python-dotenv
